Add an exact read with an optional timeout
Modified version of an exact read taken from the stk500 library.

This version spawns another process to do the read from the serial device
and blocks in receive. If the timeout is reached before the exact number
of bytes is read (the default is infinity), any accumulated buffer is
thrown away and an error is returned.