Update README
Remove some of the scary language. Hopefully most of the portability
issues have now been dealt with. The runtime lookup of termios, etc., is
not performance critical. Most applications will be doing read/writes to
the serial device. As for the safety of the native serial interface,
that makes srly useful for finding bugs in the OS!

Also remove the note about needing to set the serial device into raw
mode. There are 2 situations where blocking mode might be used:

* the caller passes in a fd opened by other means

  There's a note already explaining that the fd should be opened in
  O_NONBLOCK|O_NOCTTY mode.

* the caller uses tcsetattr/3 to change the mode